backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.3.4
postgrest==0.16.8
httpx==0.27.0
Pillow==10.1.0
